BinaryContentStorage put메서드 매개변수 수정
