+++ src/main/java/com/sprint/mission/discodeit/service/binarycontent/BinaryContentManager.java
@@ -13,6 +13,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.UUID;
@@ -26,14 +27,27 @@ public class BinaryContentManager {
     private final BinaryContentRepository binaryContentRepository;
 
     public BinaryContent saveFileAndMeta(MultipartFile file) {
-        UUID id = binaryContentStorage.put(file);
-
+        UUID id = UUID.randomUUID();
+        try {
+            binaryContentStorage.put(id, file.getBytes());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
         BinaryContent binaryContent = new BinaryContent(
                 id.toString(),
                 file.getContentType(),
                 file.getSize());
         binaryContentRepository.save(binaryContent);
         return binaryContentRepository.save(binaryContent);
+    }
+
+    private void validateContentType(String contentType) {
+        if (contentType == null ||
+                !(contentType.equals("image/png")
+                        || contentType.equals("image/jpeg")
+                        || contentType.equals("image/gif"))) {
+            throw new IllegalArgumentException("허용되지 않는 파일 형식입니다: " + contentType);
+        }
     }
