+++ src/main/java/com/sprint/mission/discodeit/service/binarycontent/LocalBinaryContentStorge.java
@@ -49,27 +49,17 @@ public class LocalBinaryContentStorge implements BinaryContentStorage {
     }
 
     @Override
-    public UUID put(MultipartFile file) {
-        String contentType = file.getContentType();
-        validateContentType(contentType);
-        String fileName = UUID.randomUUID().toString();
-        Path filePath = ROOT_PATH.resolve(fileName);
+    public UUID put(UUID uuid, byte[] bytes) {
+        Path filePath = ROOT_PATH.resolve(uuid.toString());
         try {
-            file.transferTo(filePath);
+            Files.write(filePath, bytes);
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
-        return UUID.fromString(fileName);
+        return uuid;
     }
 
-    private void validateContentType(String contentType) {
-        if (contentType == null ||
-                !(contentType.equals("image/png")
-                        || contentType.equals("image/jpeg")
-                        || contentType.equals("image/gif"))) {
-            throw new IllegalArgumentException("허용되지 않는 파일 형식입니다: " + contentType);
-        }
-    }
+
 
     @Override
     public UrlResource getUrlResource(String fileName) {
