+++ src/test/java/com/sprint/mission/discodeit/storage/AWSS3Test.java
@@ -1,0 +1,9 @@
+package com.sprint.mission.discodeit.storage;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+public class AWSS3Test {
+
+
+}
